add helper to compare topo so that ordering of topo objects does not need to be the same
