+++ tests/uio66/uio66-linker-arb-terms.lmpdat
@@ -38,8 +38,8 @@ Angle Coeffs
 Dihedral Coeffs
 
  1 dih1
- 2 dih1
- 3 dih1
+ 2 dih2
+ 3 dih3
 
 Improper Coeffs
